+++ client/src/components/common/Header.tsx
@@ -4,7 +4,7 @@ import { BiBell } from "react-icons/bi";
 import "../../styles/css/base/reset.css";
 import "../../styles/scss/components/common/header.scss";
 import { Link } from "react-router-dom";
-import NotificationAlarm from "./NotificationAlarm";
+import NotificationAlarm from "../notifications/NotificationAlarm";
 
 export const Header: React.FC = () => {
   return (
